requirements.txt
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
pydantic-settings>=2.0.0
langgraph>=0.2.0
chromadb>=0.5.0
pymupdf>=1.24.0
tiktoken>=0.7.0
unstructured>=0.14.0
python-docx>=1.1.0

# 测试依赖
pytest>=8.0.0
pytest-cov>=5.0.0
